Create search_region_function.py
This problem set covers the following topics:
  - Basics of algorithms: correctness and running time complexity.
  - Time Complexity: O, big-Omega and big-Theta Notations.
  - Proving Correctness of Algorithms through Inductive Invariants.
  - Merge Sort: Proving Correctness.
  
__Important Note__

Although this is a programming assignment, we have asked you to work on the "design" and provided opportunities for you to analyze your solution and describe your design. __However, those parts will not be graded.__ You are welcome to compare your answers against our solutions once you have completed the assignments. Our solutions are provided at the very end.